feat: rename BankRegistryEntity to BankEntity
