+++ splice-service/src/bank-registry/bank.entity.ts
