+++ splice-service/test/e2e/bank-management.e2e-spec.ts
@@ -12,8 +12,8 @@ import { AuthModule } from '../../src/auth/auth.module';
 import { BankConnectionController } from '../../src/bank-connections/bank-connection.controller';
 import { BankConnectionEntity } from '../../src/bank-connections/bank-connection.entity';
 import { BankConnectionService } from '../../src/bank-connections/bank-connection.service';
+import { BankEntity } from '../../src/bank-registry/bank.entity';
 import { BankRegistryController } from '../../src/bank-registry/bank-registry.controller';
-import { BankEntity } from '../../src/bank-registry/bank-registry.entity';
 import { BankRegistryService } from '../../src/bank-registry/bank-registry.service';
 import { UserEntity } from '../../src/users/user.entity';
 import { UsersModule } from '../../src/users/users.module';
