+++ splice-service/test/unit/bank-connections/bank-connection.service.spec.ts
@@ -5,7 +5,7 @@ import { BankConnectionStatus, BankSourceType } from '@splice/api';
 import type { Repository } from 'typeorm';
 import { BankConnectionEntity } from '../../../src/bank-connections/bank-connection.entity';
 import { BankConnectionService } from '../../../src/bank-connections/bank-connection.service';
-import { BankEntity } from '../../../src/bank-registry/bank-registry.entity';
+import { BankEntity } from '../../../src/bank-registry/bank.entity';
 import { BankRegistryService } from '../../../src/bank-registry/bank-registry.service';
 import { MOCK_USER, MOCK_USER_UUID } from '../../mocks/mocks';
